Implementação de módulo leitor
