+++ test/alunos_singleline.csv
@@ -1,0 +1,1 @@
+1,Nico,10.0,8.0,9.0,9.5
